Add a link to details on Wikipedia's bias
This links to a pre-censored version of the Wikipedia page, which summarizes its co-founder's (Larry Sanger) views on the specific biases by Wikipedia. I suppose we can provide a different link; in general, it would be good to link to somewhere where the reader can glean more information regarding the bias.